:recycle: Move repo to usecase/service
