+++ 1_usecase/service/user_service_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/Ras96/clean-architecture-sample/0_domain/model"
-	"github.com/Ras96/clean-architecture-sample/0_domain/repository"
-	"github.com/Ras96/clean-architecture-sample/0_domain/repository/mock_repository"
+	"github.com/Ras96/clean-architecture-sample/1_usecase/repository"
+	"github.com/Ras96/clean-architecture-sample/1_usecase/repository/mock_repository"
 	"github.com/Ras96/clean-architecture-sample/util/random"
 	"github.com/gofrs/uuid"
 	"github.com/golang/mock/gomock"
