+++ 2_interface/repository/user_repo.go
@@ -2,7 +2,7 @@ package repository
 
 import (
 	"github.com/Ras96/clean-architecture-sample/0_domain/model"
-	"github.com/Ras96/clean-architecture-sample/0_domain/repository"
+	"github.com/Ras96/clean-architecture-sample/1_usecase/repository"
 	"github.com/Ras96/clean-architecture-sample/2_interface/database"
 	"github.com/gofrs/uuid"
 )
